backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
python-slugify==4.0.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
validate-email==1.3
